+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-[Unreleased]## [0.5.5.2b3]
+[Unreleased]## [0.5.5.2b4]
+
+## [0.5.5.2b3]- 2023-07-05
 ### Added
 - Core01EnvironmentManager.find_project_root_directory
 - MapManager.add_bar_chart_in_map
@@ -20,10 +22,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Core01DataManager logger now uses `__name__`
 - Core01DataManager - removed deprecated logger methods
 
-
 ## [0.5.5.2b1]- 2023-05-23
 ### Changed
 - Support enable_refine_conflict in Core02OptimizationEngine and ScenarioRunner
+- Requires Pandas 2.0 which fixes an issue with SQLAlchemy 2.0 Future
 
 ## [0.5.5.2b0]- 2023-05-02
 ### Added